+++ tests/tests/test-01-mirtOne.R
@@ -58,6 +58,9 @@ test_that('dich', {
     expect_is(modm6, 'ConfirmatoryClass')
     cfs <- as.numeric(do.call(c, coef(modm6)))
     expect_equal(cfs, c(0.969, 1.851, 0, 1, 1.074, 0.808, 0, 1, 1.717, 1.811, 0, 1, 0.763, 0.486, 0, 1, 0.731, 1.852, 0, 1, 0, 1), tolerance = 1e-2)
+    modideal <- mirt(data, 1, verbose=FALSE, itemtype='ideal')
+    cfs <- as.numeric(do.call(c, coef(modideal)))
+    expect_equal(cfs, c(0.288, -0.568, 0.422, -0.891, 0.57, -0.564, 0.292, -1, 0.207, -0.559, 0, 1), tolerance = 1e-2)
 
     fm1 <- fscores(modm1, verbose = FALSE)
     expect_is(fm1, 'matrix')